move the `insert_activity` method
